// frontend/StockManager/generated-sources/openapi/models/WatchListDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * StockManager API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { WatchListItemDto } from './WatchListItemDto';
import {
    WatchListItemDtoFromJSON,
    WatchListItemDtoFromJSONTyped,
    WatchListItemDtoToJSON,
    WatchListItemDtoToJSONTyped,
} from './WatchListItemDto';

/**
 * 
 * @export
 * @interface WatchListDto
 */
export interface WatchListDto {
    /**
     * 
     * @type {number}
     * @memberof WatchListDto
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof WatchListDto
     */
    userId?: number;
    /**
     * 
     * @type {Array<WatchListItemDto>}
     * @memberof WatchListDto
     */
    watchListItems?: Array<WatchListItemDto> | null;
}

/**
 * Check if a given object implements the WatchListDto interface.
 */
export function instanceOfWatchListDto(value: object): value is WatchListDto {
    return true;
}

export function WatchListDtoFromJSON(json: any): WatchListDto {
    return WatchListDtoFromJSONTyped(json, false);
}

export function WatchListDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): WatchListDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'userId': json['userId'] == null ? undefined : json['userId'],
        'watchListItems': json['watchListItems'] == null ? undefined : ((json['watchListItems'] as Array<any>).map(WatchListItemDtoFromJSON)),
    };
}

export function WatchListDtoToJSON(json: any): WatchListDto {
    return WatchListDtoToJSONTyped(json, false);
}

export function WatchListDtoToJSONTyped(value?: WatchListDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'userId': value['userId'],
        'watchListItems': value['watchListItems'] == null ? undefined : ((value['watchListItems'] as Array<any>).map(WatchListItemDtoToJSON)),
    };
}