// frontend/StockManager/generated-sources/openapi/models/WatchListItemDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * StockManager API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface WatchListItemDto
 */
export interface WatchListItemDto {
    /**
     * 
     * @type {number}
     * @memberof WatchListItemDto
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof WatchListItemDto
     */
    stockId?: number;
    /**
     * 
     * @type {number}
     * @memberof WatchListItemDto
     */
    watchListId?: number;
    /**
     * 
     * @type {number}
     * @memberof WatchListItemDto
     */
    entryPrice?: number | null;
    /**
     * 
     * @type {string}
     * @memberof WatchListItemDto
     */
    note?: string | null;
}

/**
 * Check if a given object implements the WatchListItemDto interface.
 */
export function instanceOfWatchListItemDto(value: object): value is WatchListItemDto {
    return true;
}

export function WatchListItemDtoFromJSON(json: any): WatchListItemDto {
    return WatchListItemDtoFromJSONTyped(json, false);
}

export function WatchListItemDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): WatchListItemDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'stockId': json['stockId'] == null ? undefined : json['stockId'],
        'watchListId': json['watchListId'] == null ? undefined : json['watchListId'],
        'entryPrice': json['entryPrice'] == null ? undefined : json['entryPrice'],
        'note': json['note'] == null ? undefined : json['note'],
    };
}

export function WatchListItemDtoToJSON(json: any): WatchListItemDto {
    return WatchListItemDtoToJSONTyped(json, false);
}

export function WatchListItemDtoToJSONTyped(value?: WatchListItemDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'stockId': value['stockId'],
        'watchListId': value['watchListId'],
        'entryPrice': value['entryPrice'],
        'note': value['note'],
    };
}