// frontend/StockManager/generated-sources/openapi/models/WatchListItemUpdateDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * StockManager API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface WatchListItemUpdateDto
 */
export interface WatchListItemUpdateDto {
    /**
     * 
     * @type {number}
     * @memberof WatchListItemUpdateDto
     */
    stockId?: number;
    /**
     * 
     * @type {number}
     * @memberof WatchListItemUpdateDto
     */
    watchListId?: number;
    /**
     * 
     * @type {number}
     * @memberof WatchListItemUpdateDto
     */
    entryPrice?: number;
    /**
     * 
     * @type {string}
     * @memberof WatchListItemUpdateDto
     */
    note?: string | null;
}

/**
 * Check if a given object implements the WatchListItemUpdateDto interface.
 */
export function instanceOfWatchListItemUpdateDto(value: object): value is WatchListItemUpdateDto {
    return true;
}

export function WatchListItemUpdateDtoFromJSON(json: any): WatchListItemUpdateDto {
    return WatchListItemUpdateDtoFromJSONTyped(json, false);
}

export function WatchListItemUpdateDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): WatchListItemUpdateDto {
    if (json == null) {
        return json;
    }
    return {
        
        'stockId': json['stockId'] == null ? undefined : json['stockId'],
        'watchListId': json['watchListId'] == null ? undefined : json['watchListId'],
        'entryPrice': json['entryPrice'] == null ? undefined : json['entryPrice'],
        'note': json['note'] == null ? undefined : json['note'],
    };
}

export function WatchListItemUpdateDtoToJSON(json: any): WatchListItemUpdateDto {
    return WatchListItemUpdateDtoToJSONTyped(json, false);
}

export function WatchListItemUpdateDtoToJSONTyped(value?: WatchListItemUpdateDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'stockId': value['stockId'],
        'watchListId': value['watchListId'],
        'entryPrice': value['entryPrice'],
        'note': value['note'],
    };
}